backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0
